requirements.txt
aiomysql==0.1.1
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.89.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
numpy==1.24.2
pydantic==1.10.4
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
pytz
python-dateutil
